修改指数越界问题，诊断页拉长问题

+++ StockDoctor/StockDoctor/DetailController/LineVIews/KLineViews/KLineView.m
@@ -22,7 +22,7 @@
     float _topPrice;
     float _buttomPrice;
     float _halfPrice;
-    long _topVolume;
+    long long _topVolume;
 }
 
 - (id)initWithCoder:(NSCoder *)aDecoder
@@ -34,9 +34,6 @@
     }
     return self;
 }
-
-
-
 
 - (void)reloadData:(KLineModel *)model
 {
